fix file exitst for symlink unlink

+++ src/bp-document/bp-document-functions.php
@@ -4584,7 +4584,7 @@ function bb_document_delete_older_symlinks() {
 			continue;
 		}
 		$file = $dir . '/' . $file;
-		if ( filemtime( $file ) < $limit ) {
+		if ( file_exists( $file ) && filemtime( $file ) < $limit ) {
 			$list[] = $file;
 			unlink( $file );
 		}
