+++ src/bp-media/bp-media-functions.php
@@ -3804,7 +3804,7 @@ function bb_media_delete_older_symlinks() {
 			continue;
 		}
 		$file = $dir . '/' . $file;
-		if ( filemtime( $file ) < $limit ) {
+		if ( file_exists( $file ) && filemtime( $file ) < $limit ) {
 			$list[] = $file;
 			unlink( $file );
 		}
